Fix oauth and checkUserTokens bug

+++ src/scripts/oauth/ClearApplicationData.js
@@ -1,7 +1,15 @@
 
 // Log user out on request
-export async function ClearApplicationData() {
+// @deleteDefinitions {boolean}, @authorizeAgain {boolean}
+export async function ClearApplicationData(deleteDefinitions = true, authorizeAgain = false) {
     window.localStorage.clear();
     window.sessionStorage.clear();
-    indexedDB.deleteDatabase('keyval-store');
+
+    if (deleteDefinitions) {
+        indexedDB.deleteDatabase('keyval-store');
+    };
+
+    if (authorizeAgain) {
+        window.location.href = `https://www.bungie.net/en/oauth/authorize?&client_id=${import.meta.env.CLIENT_ID}&response_type=code`;
+    };
 };
